requirements.txt
icecream==2.1.3
iexfinance==0.5.0
numpy==1.26.4
pandas==2.2.2
plotly==5.18.0
pyspark==3.5.1
python-dotenv==1.0.1
streamlit==1.34.0
